requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg[binary]
django-filter==24.1
drf-spectacular==0.27.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
gunicorn==21.2.0
